# Docs/OpenAPI.json
{
  "produces": ["application/json"],
  "schemes": ["http"],
  "swagger": "2.0",
  "info": {
    "description": "Gateway service for Classroom microservices. Handles authentication, routing, and request aggregation.",
    "title": "Classroom Gateway API",
    "contact": {},
    "version": "1.0"
  },
  "host": "127.0.0.1",
  "basePath": "/api",
  "paths": {
...
    "/tasks/create": {
      "post": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Создает новую задачу в курсе",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["Tasks"],
        "summary": "Создание задачи",
        "parameters": [
          {
            "description": "Данные для создания задачи",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateTaskRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/CreateTaskResponse"
            }
          },
          "400": {
            "description": "Некорректные данные",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "401": {
            "description": "Требуется авторизация",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "403": {
            "description": "Доступ запрещён",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Внутренняя ошибка",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "503": {
            "description": "Сервис недоступен",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/tasks/task": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Возвращает детальную информацию о задаче в рамках курса",
        "produces": ["application/json"],
        "tags": ["Tasks"],
        "summary": "Получение задачи",
        "parameters": [
          {
            "type": "string",
            "example": "\"5a430d16-851d-45a9-b55b-15838785adea\"",
            "description": "ID задачи",
            "name": "task_id",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/GetTaskResponse"
            }
          },
          "400": {
            "description": "Некорректные данные",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "401": {
            "description": "Требуется авторизация",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "403": {
            "description": "Доступ запрещен",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "404": {
            "description": "Задача не найдена",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервера",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "503": {
            "description": "Сервис недоступен",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/tasks/student-statuses": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Возвращает статусы выполнения задачи для студентов",
        "produces": ["application/json"],
        "tags": ["Tasks"],
        "summary": "Получение статусов студентов",
        "parameters": [
          {
            "type": "string",
            "example": "\"21dad0c3-dcea-4c19-b501-fb2fe888f683\"",
            "description": "ID задачи",
            "name": "task_id",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/GetStudentStatusesResponse"
            }
          },
          "400": {
            "description": "Некорректные данные",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "401": {
            "description": "Требуется авторизация",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "403": {
            "description": "Доступ запрещен",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "404": {
            "description": "Данные не найдены",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервера",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "503": {
            "description": "Сервис недоступен",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/tasks/task/changestatus": {
      "patch": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Обновляет статус выполнения задачи студентом",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["Tasks"],
        "summary": "Изменение статуса задачи",
        "parameters": [
          {
            "description": "Данные для изменения статуса",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ChangeStatusTaskRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ChangeStatusTaskResponse"
            }
          },
          "400": {
            "description": "Некорректные данные",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "401": {
            "description": "Требуется авторизация",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "404": {
            "description": "Задача не найдена",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервера",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "503": {
            "description": "Сервис недоступен",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/tasks/task/update": {
      "put": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Обновляет информацию о задаче",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["Tasks"],
        "summary": "Обновление задачи",
        "parameters": [
          {
            "description": "Данные для обновления",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdateTaskRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/UpdateTaskResponse"
            }
          },
          "400": {
            "description": "Некорректные данные",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "401": {
            "description": "Требуется авторизация",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "404": {
            "description": "Задача не найдена",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервера",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "503": {
            "description": "Сервис недоступен",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/tasks/task/delete": {
      "delete": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Удаляет задачу по идентификатору",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["Tasks"],
        "summary": "Удаление задачи",
        "parameters": [
          {
            "description": "Идентификатор задачи",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/DeleteTaskRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/DeleteTaskResponse"
            }
          },
          "400": {
            "description": "Некорректные данные",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "401": {
            "description": "Требуется авторизация",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "404": {
            "description": "Задача не найдена",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервера",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "503": {
            "description": "Сервис недоступен",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/tasks/student-tasks": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Возвращает список задач для конкретного студента",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["Tasks"],
        "summary": "Получение задач студента",
        "parameters": [
          {
            "type": "string",
            "example": "44e7f029-82cc-46f5-83e8-34b7d056ce32",
            "description": "Идентификатор курса",
            "name": "course_id",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/GetTasksForStudentResponse"
            }
          },
          "400": {
            "description": "Некорректные данные",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "401": {
            "description": "Требуется авторизация",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "404": {
            "description": "Студент не найден",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервера",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "503": {
            "description": "Сервис недоступен",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/tasks/teacher-tasks": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Возвращает список задач курса для учителя",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["Tasks"],
        "summary": "Получение списка задач для учителя",
        "parameters": [
          {
            "type": "string",
            "example": "44e7f029-82cc-46f5-83e8-34b7d056ce32",
            "description": "Идентификатор курса",
            "name": "course_id",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/GetTasksResponse"
            }
          },
          "400": {
            "description": "Некорректные данные",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "401": {
            "description": "Требуется авторизация",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "404": {
            "description": "Курс не найден",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервера",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "503": {
            "description": "Сервис недоступен",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    }
  },
  "definitions": {
...
    "ChangeStatusTaskRequest": {
      "description": "Позволяет изменить статус выполнения задания",
      "type": "object",
      "properties": {
        "task_id": {
          "description": "ID задания",
          "type": "string",
          "x-order": "0",
          "example": "d277084b-e1f6-4670-825b-53951d20b5d3"
        }
      }
    },
    "ChangeStatusTaskResponse": {
      "description": "Пустой ответ при успешном изменении",
      "type": "object"
    },
...
    "CreateTaskRequest": {
      "description": "Параметры для создания нового задания в курсе",
      "type": "object",
      "properties": {
        "course_id": {
          "description": "ID курса",
          "type": "string",
          "x-order": "0",
          "example": "d277084b-e1f6-4670-825b-53951d20b5d3"
        },
        "title": {
          "description": "Название задания",
          "type": "string",
          "x-order": "1",
          "example": "Лабораторная работа 1"
        },
        "description": {
          "description": "Описание задания",
          "type": "string",
          "x-order": "2",
          "example": "Реализовать алгоритм сортировки"
        }
      }
    },
    "CreateTaskResponse": {
      "description": "Возвращает ID созданного задания",
      "type": "object",
      "properties": {
        "task_id": {
          "description": "ID созданного задания",
          "type": "string",
          "x-order": "0",
          "example": "d277084b-e1f6-4670-825b-53951d20b5d3"
        }
      }
    },
...
    "DeleteTaskRequest": {
      "description": "Требует ID курса и задания для удаления",
      "type": "object",
      "properties": {
        "task_id": {
          "description": "ID задания",
          "type": "string",
          "x-order": "0",
          "example": "d277084b-e1f6-4670-825b-53951d20b5d3"
        }
      }
    },
    "DeleteTaskResponse": {
      "description": "Пустой ответ при успешном удалении",
      "type": "object"
    },
...
    "ErrorResponse": {
      "description": "Cтандартизированный формат всех ошибок в API.",
      "type": "object",
      "properties": {
        "code": {
          "description": "Код ошибки (HTTP статус код)",
          "type": "integer",
          "example": 400
        },
        "message": {
          "description": "Сообщение об ошибке",
          "type": "string",
          "example": "Bad request"
        }
      }
    },
...
    "GetStudentStatusesResponse": {
      "description": "Содержит статусы выполнения задания студентами",
      "type": "object",
      "properties": {
        "statuses": {
          "description": "Массив статусов",
          "type": "array",
          "items": {
            "$ref": "#/definitions/TaskStatus"
          },
          "x-order": "0"
        }
      }
    },
    "GetTaskResponse": {
      "description": "Возвращает полные данные задания",
      "type": "object",
      "properties": {
        "task": {
          "description": "Объект задания",
          "allOf": [
            {
              "$ref": "#/definitions/Task"
            }
          ],
          "x-order": "0"
        }
      }
    },
    "GetTasksForStudentResponse": {
      "description": "Содержит задания с информацией о статусе выполнения",
      "type": "object",
      "properties": {
        "tasks": {
          "description": "Массив заданий со статусами",
          "type": "array",
          "items": {
            "$ref": "#/definitions/StudentTask"
          },
          "x-order": "0"
        }
      }
    },
    "GetTasksResponse": {
      "description": "Содержит массив заданий в курсе",
      "type": "object",
      "properties": {
        "tasks": {
          "description": "Массив заданий",
          "type": "array",
          "items": {
            "$ref": "#/definitions/Task"
          },
          "x-order": "0"
        }
      }
    },
...
    "UpdateTaskRequest": {
      "description": "Позволяет обновить данные задания",
      "type": "object",
      "properties": {
        "task_id": {
          "description": "ID задания",
          "type": "string",
          "x-order": "0",
          "example": "d277084b-e1f6-4670-825b-53951d20b5d3"
        },
        "title": {
          "description": "Новое название (опционально)",
          "type": "string",
          "x-order": "1",
          "example": "Обновленное название"
        },
        "description": {
          "description": "Новое описание (опционально)",
          "type": "string",
          "x-order": "2",
          "example": "Новые требования"
        }
      }
    },
    "UpdateTaskResponse": {
      "description": "Пустой ответ при успешном обновлении",
      "type": "object"
    }
  },
  "securityDefinitions": {
    "BearerAuth": {
      "description": "Enter JWT token in format: \"Bearer {token}\"",
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}